FIR IDE: quickfix to remove branch for SENSELESS_NULL_IN_WHEN
Diagnostic REDUNDANT_ELSE_IN_WHEN is not reported by FIR as for now.
