+++ plugins/kotlin/idea/tests/testData/quickfix/when/removeRedundantBranch.kt
@@ -7,4 +7,3 @@ fun test(x: Int): String {
         else -> ""
     }
 }
-/* IGNORE_FIR */
