+++ plugins/kotlin/idea/tests/testData/quickfix/when/removeRedundantBranch.kt.after
@@ -6,4 +6,3 @@ fun test(x: Int): String {
         else -> ""
     }
 }
-/* IGNORE_FIR */
